+++ R/export.sc_pet.R
@@ -17,26 +17,27 @@ export.sc_pet <- function(object,
   if (is.na(footnote)) {
     if (object$decreasing) {
       footnote <- c(
-        "Assumed decreasing values in the B-phase.",
+        "Assumed decreasing values in the B-phase",
         "Binomial test alternative hypothesis: true probability < 50%",
-        sprintf(
-          "PET CI: Percent of values less than the lower %d%% confidence threshold", 
-          object$ci.percent
-        )
+        "Single-sided test"
       )
     } else {
       footnote <- c(
+        "Assumed increasing values in the B-phase",
         "Binomial test alternative hypothesis: true probability > 50%",
-        sprintf(
-          "PET CI: Percent of values greater than the upper %d%% confidence threshold", 
-          object$ci.percent
-        )
+        "Single-sided test"
       )
     }
   }
   
   object$PET$binom.p <- .nice_p(object$PET$binom.p)
   names(object$PET)[4] <- "p (binomial test)"
+  
+  new <- paste0(
+    "Percentage ", if (object$decreasing) "< lower " else "> upper ", 
+    object$ci * 100, " CI"
+  )
+  names(object$PET)[3] <- new
   object$PET <- round_numeric(object$PET, round)
   
   table <- .create_table(