+++ inst/shiny/server.R
@@ -41,8 +41,12 @@ server <- function(input, output, session) {
         cat(paste0("loaded example ", input$scdf_example))
       )
       output$load_output <- renderPrint({
-        do.call("summary", list(my_scdf()))
+         do.call("summary", list(my_scdf()))
       })
+      #output$load_html_output <- renderUI({
+      #   x <- do.call("summary", list(my_scdf()))
+      #   export(x) |> HTML()
+      #})
       output$scdf_messages <- renderPrint(cat(""))
     } else {
       my_scdf(NULL)
@@ -671,10 +675,24 @@ server <- function(input, output, session) {
     )  
     
     ci <- input$pt_ci
-   
+    
+    pt_method <- deparse(as.list(input$pt_method))
+    if (input$pt_method_user != "") {
+      
+      if (length(input$pt_method) >0) {
+        pt_method <- paste0("\"", input$pt_method, "\"",collapse = ",")
+        pt_method <- paste0(
+          "list(user = function(scdf) {", input$pt_method_user,"}, ",
+          pt_method, ")"
+        )
+      } else {
+        pt_method <- paste0("list(user = function(scdf) {", input$pt_method_user,"})")
+      }
+    }
+    
     syntax <- paste0(
       syntax, res$pipe,
-      "  power_test(method = ", deparse(input$pt_method), ", ",
+      "  power_test(method = ", pt_method, ", ",
       "effect = ", deparse(input$pt_effect), ", ",
       "n_sim = ", input$pt_n, 
       if (!identical(ci, "")) paste0(", ci = ", ci),