+++ NEWS.md
@@ -1,12 +1,41 @@
 # scan 0.61.999
 
-## New
-
-## Changes
+## New functions / features
 
 - `tau_u()`: New method `"tarlow"` calculates Tau-U as implemented in an R code and online calculator by Tarlow (2017). Here, tau values are calculated as in the `method = "complete", continuity_correction = TRUE, tau_method = "a"`. Inferential statistics are calculated based an tau b and the standard deviation for S is derived directly from Kendall's Tau B analysis (different from the `parker` and `complete` methods). 
-- `tau_u()`: Method `"parker"` ignores the `tau_method` setting and sets `continuity_correction = FALSE`. This follows the Parker (2011) paper. There, the inferential statistics are calculated using Kendall's Tau b while the actual Tau calculation applies Kendall's Tau a (without ties).
+- `rand_test()`: It is now possible to provide new functions for calculating the statistic directly with a list to the `statistic_function` argument. This list must have an element named `statistic` with a function that takes two arguments `a` and `b` and returns a single numeric value. A second element of the list is named `aggregate` which takes a function with one numeric argument that returns a numeric argument. This function is used to aggregate the values of a multiple case design. If you do not provide this element, it uses the default `function(x) sum(x)/length(x)`. The third optional argument is `name` which provides a name for your user function.
 
+```r
+userstat <- list(
+  statistic = function(a, b) median(b) - median(a), 
+  aggregate = function(x) median(x),
+  name = "median B - A"
+)
+
+rand_test(exampleAB, statistic_function = userstat , complete = TRUE)
+
+# which is identical to:
+rand_test(exampleAB, statistic = "Median B-A" , complete = TRUE)
+```
+
+- `rand_test()`: Returns startpoints for the random permutations.
+- `plot_rand()`: New argument `type` when `"xy"` a plot with splitpoints and statistics is drawn. This allows to see graphically at which measurement time a statistic changes.
+
+```r
+Leidig2018[4] |> 
+  na.omit() |> 
+  rand_test(complete = TRUE, limit = 1, statistic = "SMD glass") |> 
+  plot_rand(type = "xy")
+```
+
+
+- `na.omit.scdf()`: scdf method for generic `na.omit()`. Removes any row with a missing value from an scdf.
+
+
+## Corrections / Changes 
+
+- `scdf()`: Throws an error when argument `phase_starts` is set and the beginning of the first phase is not the first measurement.
+- `tau_u()`: Method `"parker"` ignores the `tau_method` setting and sets `continuity_correction = FALSE`. This follows the Parker (2011) paper. There, the inferential statistics are calculated using Kendall's Tau b while the actual Tau calculation applies Kendall's Tau a (without ties).
 
 
 ## Solved bugs